requirements.txt
langchain==0.3.7
langchain-groq==0.2.3
langchain-community==0.3.7
langchain-huggingface==0.1.2
chromadb==0.5.3
sentence-transformers==3.3.1
streamlit==1.40.0
ragas==0.2.5
langsmith==0.1.147
python-dotenv==1.0.1
pandas==2.2.0
datasets==3.1.0
